change the src/ structure and update the dashboard based on reflection
